+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Add a base template for listing objects (`__object_list.html`)
 * Show filter form in list views in a collapsible accordion
 * Show an alert on the `WorkspaceSharingAuditReview` and `ManagedGroupMembershipAuditReview` views if an error is found in the audit result for the workspace or group, respectively. Such errors could indicate that the sharing or membership audits could be incorrect.
+* Add a new mixin for use with workspace adapters (`WorkspaceSharingAdapterMixin`) that will automatically share a workspace upon creation or import with a set of specified groups
+* Break advanced docs into its own TOC and sub-section source documents
 
 ## 0.33.0 (2025-08-04)